Support Strata Source expanded limits
Strata Source (BSP version 25) increases many object count limits,
which necessitated upgrading many lump fields with larger data types
(many short to int upgrades). It also increases the maximum map extents
to 65536 units from the origin.

This commit was re-formed from the original found here:
https://github.com/StrataSource/bspsrc/commit/ecc46a2a8ca2c4e81fb281241c937e290402a2b7

+++ bspsrc-decompiler/src/main/java/info/ata4/bspsrc/decompiler/modules/geom/BrushSideFaceMapper.java
@@ -181,10 +181,10 @@ public class BrushSideFaceMapper extends ModuleRead {
 
 	private static class FaceIndexKey {
 		public final int pnum;
-		public final short texinfo;
-		public final short dispInfo;
+		public final int texinfo;
+		public final int dispInfo;
 
-		private FaceIndexKey(int pnum, short texinfo, short dispInfo) {
+		private FaceIndexKey(int pnum, int texinfo, int dispInfo) {
 			this.pnum = pnum;
 			this.texinfo = texinfo;
 			this.dispInfo = dispInfo;
