+++ bspsrc-lib/src/main/java/info/ata4/bspsrc/lib/struct/DFaceBSP18.java
@@ -65,9 +65,9 @@ public class DFaceBSP18 extends DFace {
         out.writeByte(side);
         out.writeByte(onnode);
         out.writeInt(fstedge);
-        out.writeShort(numedge);
-        out.writeShort(texinfo);
-        out.writeShort(dispInfo);
+        out.writeShort((short)numedge);
+        out.writeShort((short)texinfo);
+        out.writeShort((short)dispInfo);
         out.writeUnsignedShort(surfaceFogVolumeID);
         out.writeBytes(styles);
         out.writeInt(lightofs);
